complete Add to Cart Part

+++ app/views/pages/cart.php
@@ -37,34 +37,42 @@
                         $cartTotal = 0.00;
                         
                         if (!empty($cartItems)) {
+                           
                             foreach ($cartItems as $cartItem) {
-                                $cartTotal += ($cartItem->quantity * $cartItem->unitPrice);
+                                
+                                if (is_array($cartItem)) {
+                                    $cartTotal += ($cartItem['quantity'] * $cartItem['unitPrice']);
+                                }
+                                
+                                // $cartTotal += ($cartItem['quantity'] * $cartItem['unitPrice']);
                             }
                         }
                     ?>
               
                     <?php if (!empty($cartItems)) : ?>
                         <?php foreach ($cartItems as $item) : ?>
+                            <?php if (is_array($cartItem)) : ?>
                             
-                            <tr class="cartItm" data-item-id="<?php echo $item['itemId']; ?>">
-                                <td>
-                                    <div class="cart-info">
-                                        
-                                        <img src="<?php echo URLROOT?>/public/images/products/Coconut-APM-D-1.png" alt="<?php echo $item['itemName']; ?>">
-                                        <div>
-                                            <h4><?php echo $item['itemName']; ?></h4>
-                                            <small>Price: LKR <?php echo number_format($item['unitPrice'], 2); ?></small><br>
-                                            <!-- Add a "Remove" link with an onclick event to trigger the popup message -->
-                                                <a  onclick="showRemoveConfirmation('<?php echo $item['itemId']; ?>' )" href="<?php echo URLROOT ?>/Cart/deleteItem?itemId=<?php echo $item['itemId']; ?>&uId=<?php echo $item['uId']; ?>">Remove</a>                                         
+                                <tr class="cartItm" data-item-id="<?php echo $item['itemId']; ?>">
+                                    <td>
+                                        <div class="cart-info">
+                                            
+                                            <img src="<?php echo URLROOT?>/public/images/products/Coconut-APM-D-1.png" alt="<?php echo $item['itemName']; ?>">
+                                            <div>
+                                                <h4><?php echo $item['itemName']; ?></h4>
+                                                <small>Price: LKR <?php echo number_format($item['unitPrice'], 2); ?></small><br>
+                                                <!-- Add a "Remove" link with an onclick event to trigger the popup message -->
+                                                    <a  onclick="showRemoveConfirmation('<?php echo $item['itemId']; ?>' )" href="<?php echo URLROOT ?>/Cart/deleteItem?itemId=<?php echo $item['itemId']; ?>&uId=<?php echo $item['uId']; ?>">Remove</a>                                         
+                                            </div>
                                         </div>
-                                    </div>
-                                </td>
-                                <td>
-                                    <input type="number" min=1 class="quantity-input" name="quantitiesTo[]" data-item-id="<?php echo $item['itemId']; ?>" data-item-uprice="<?php echo $item['unitPrice']; ?>"  value="<?php echo $item['quantity']; ?>">      
-                                    <input type="hidden" name="itemIds[]" value="<?php echo $item['itemId']; ?>">                                   
-                                </td>
-                                <td class="subtotal" data-item-id="<?php echo $item['itemId']; ?>">LKR <span class="subtotal-value"><?php echo number_format($item['unitPrice'] * $item['quantity'], 2); ?></span></td>                                
-                            </tr>                           
+                                    </td>
+                                    <td>
+                                        <input type="number" min=1 class="quantity-input" name="quantitiesTo[]" data-item-id="<?php echo $item['itemId']; ?>" data-item-uprice="<?php echo $item['unitPrice']; ?>"  value="<?php echo $item['quantity']; ?>">      
+                                        <input type="hidden" name="itemIds[]" value="<?php echo $item['itemId']; ?>">                                   
+                                    </td>
+                                    <td class="subtotal" data-item-id="<?php echo $item['itemId']; ?>">LKR <span class="subtotal-value"><?php echo number_format($item['unitPrice'] * $item['quantity'], 2); ?></span></td>                                
+                                </tr>   
+                            <?php endif; ?>                        
                         <?php endforeach; ?>
                     <?php else : ?>
                     <tr>
@@ -96,7 +104,7 @@
             </table>
     </section>
 </div> 
-
+                    
 
 
 <script>
